% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiTSC.R
\name{bitsc}
\alias{bitsc}
\title{BiTSC_v1}
\usage{
bitsc(rho_0, niter_0, urlfile1, urlfile2, urlfile3)
}
\arguments{
\item{rho_0}{the porportion to be sampled}

\item{niter_0}{number of iterations}
}
\value{
the same x
}
\description{
First version of BiTSC; developed Dec 10, 2022
}
