% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiTSC_v1.R
\name{bitsc_v1}
\alias{bitsc_v1}
\title{BiTSC_v1}
\usage{
bitsc_v1(rho_0, niter_0)
}
\arguments{
\item{rho_0}{the porportion to be sampled}

\item{niter_0}{number of iterations}
}
\value{
the same x
}
\description{
First version of BiTSC; developed Dec 10, 2022
}
